+++ CHANGELOG.md
@@ -2,7 +2,19 @@
 
 ### New modules
 
+- bedops/bedmap (#37)
+- bedtools/map (#37)
+- bedtools/merge (#37)
+- bedtools/sort (#37)
+- cat/cat (#37)
+- cat/fastq (#37)
+- custom/combinepeakcounts (#37)
+- custom/consensuspeaks (#37)
+- custom/normalizepeaks (#37)
+
 ### New subworkflows
+
+- consensus_peaks (#37)
 
 ## nf-modules 0.1.0
 
@@ -24,4 +36,4 @@ Our documentation website is now live: <https://ccbr.github.io/nf-modules/> (#16
 
 ### New subworkflows
 
-- custom/filter_blacklist (#17,#27)
+- filter_blacklist (#17,#27)